Feat: Day 5 Solutions
In fact I completed day 5 on the day but didn't commit until day 6. I
definitely want to refactor the solution quite a bit as I couldn't find
quite the right abstraction for the problem. I would also like to try to
find a more functional approach that doesn't relay on mutating the
memory in place.